Rename db.ts to importDb.ts
